+++ src/orbit.lua
@@ -1,13 +1,14 @@
 local app = require("orbit.app")
 
 local orbit = {
-  _VERSION = "0.1.0",
+  _VERSION = "0.2.0",
   new = app.new,
   App = app.App,
   Context = app.Context,
   json = require("orbit.json"),
   request = require("orbit.request"),
   response = require("orbit.response"),
+  static = require("orbit.static"),
 }
 
 return orbit